FEV-355 fix PR comments
